+++ LPGSavings/Models/DefaultIntroValues.cs
@@ -5,7 +5,7 @@ namespace LPGSavings.Models
 {
     public static class DefaultIntroValues
     {
-        public const decimal LPG_PRICE = 2.56M;
-        public const decimal PB_PRICE = 5.7M;
+        public const float LPG_PRICE = 2.56f;
+        public const float PB_PRICE = 5.7f;
     }
 }
